requirements.txt
flask>=1.0.2
flask_socketio>=3.0.0
flask_mongoengine>=0.9.5
flask_login>=0.4.1
flask_mail>=0.9.1
flask_bootstrap>=3.3.7.1
flask_classy>=0.6.10
validate_email>=1.3
requests>=2.11.1
wtforms>=2.2.1
python-dateutil>=2.7.3
Flask-Babel>=0.11.2